Added Assignment 10.1
Assignment 10.1 had me making a program that would take data from a user (Name, Address and Phone Number) and write said data onto a file separated by commas. I have the program write the data to a text file, but it works with word and excel with a simple tweak to lines 18, 21 and 22 (just change ".txt" to whatever file type you would like).